Move override tests to base test suite
Move override tests to base test suite. Tests that are DCB-specific should have an assumption to cover them.

#4291

(cherry picked from commit 09d19fead9cd3791e578598a3bfb82349b7ffb51)
