+++ eventsourcing/src/test/java/org/axonframework/eventsourcing/eventstore/inmemory/InMemoryStorageEngineBackedEventStoreTest.java
@@ -16,14 +16,14 @@
 
 package org.axonframework.eventsourcing.eventstore.inmemory;
 
-import org.jspecify.annotations.NonNull;
 import org.axonframework.eventsourcing.eventstore.StorageEngineBackedEventStore;
-import org.axonframework.eventsourcing.eventstore.DcbBasedStorageEngineBackedEventStoreTestSuite;
+import org.axonframework.eventsourcing.eventstore.StorageEngineBackedEventStoreTestSuite;
 import org.axonframework.messaging.core.EmptyApplicationContext;
 import org.axonframework.messaging.core.unitofwork.SimpleUnitOfWorkFactory;
 import org.axonframework.messaging.core.unitofwork.UnitOfWork;
 import org.axonframework.messaging.core.unitofwork.UnitOfWorkFactory;
 import org.axonframework.messaging.eventhandling.conversion.EventConverter;
+import org.jspecify.annotations.NonNull;
 import org.junit.jupiter.api.*;
 
 /**
@@ -33,7 +33,7 @@ import org.junit.jupiter.api.*;
  * @author John Hendrikx
  */
 class InMemoryStorageEngineBackedEventStoreTest
-        extends DcbBasedStorageEngineBackedEventStoreTestSuite<InMemoryEventStorageEngine> {
+        extends StorageEngineBackedEventStoreTestSuite<InMemoryEventStorageEngine> {
 
     private static final UnitOfWorkFactory FACTORY = new SimpleUnitOfWorkFactory(EmptyApplicationContext.INSTANCE);
