+++ extensions/spring/spring-boot-autoconfigure/src/test/java/org/axonframework/extension/springboot/eventsourcing/eventstore/jpa/AggregateBasedJpaStorageEngineBackedEventStoreIT.java
@@ -23,7 +23,7 @@ import org.axonframework.common.jdbc.PersistenceExceptionResolver;
 import org.axonframework.common.jpa.EntityManagerProvider;
 import org.axonframework.common.jpa.FactoryBasedEntityManagerProvider;
 import org.axonframework.common.jpa.SimpleEntityManagerProvider;
-import org.axonframework.eventsourcing.eventstore.AggregateBasedStorageEngineBackedEventStoreTestSuite;
+import org.axonframework.eventsourcing.eventstore.StorageEngineBackedEventStoreTestSuite;
 import org.axonframework.eventsourcing.eventstore.jpa.AggregateBasedJpaEventStorageEngine;
 import org.axonframework.eventsourcing.eventstore.jpa.JpaPollingEventCoordinator;
 import org.axonframework.extension.spring.messaging.unitofwork.SpringTransactionManager;
@@ -35,6 +35,7 @@ import org.axonframework.messaging.core.unitofwork.TransactionalUnitOfWorkFactor
 import org.axonframework.messaging.core.unitofwork.UnitOfWork;
 import org.axonframework.messaging.core.unitofwork.transaction.jpa.JpaTransactionalExecutorProvider;
 import org.axonframework.messaging.eventhandling.conversion.EventConverter;
+import org.hibernate.exception.ConstraintViolationException;
 import org.jspecify.annotations.NonNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
@@ -57,7 +58,8 @@ import javax.sql.DataSource;
  */
 @SpringBootTest(classes = TestConfig.class)
 @ImportAutoConfiguration(JpaTransactionAutoConfiguration.class)
-class AggregateBasedJpaStorageEngineBackedEventStoreIT extends AggregateBasedStorageEngineBackedEventStoreTestSuite<AggregateBasedJpaEventStorageEngine> {
+class AggregateBasedJpaStorageEngineBackedEventStoreIT extends
+        StorageEngineBackedEventStoreTestSuite<AggregateBasedJpaEventStorageEngine> {
 
     private static AggregateBasedJpaEventStorageEngine engine;
